unit-tests/trivial: move server.go test to the new server_test.go file
